Add Elastic logo to UI Framework doc site. (#11686)
