+++ ui_framework/doc_site/src/components/guide_code_viewer/_guide_code_viewer.scss
@@ -13,7 +13,7 @@
 
   @include scrollbar;
 
-  @include whenNarrowerThan($normalBreakpoint) {
+  @include whenNarrowerThan($guideNormalBreakpoint) {
     width: $guideCodeViewerSmallWidth;
   }
