+++ ui_framework/doc_site/src/components/guide_page/_guide_page.scss
@@ -22,7 +22,7 @@
     width: calc(100vw - #{$leftMargin + $rightMargin + $scrollBarWidth + $guideCodeViewerWidth});
   }
 
-  @include whenNarrowerThan($normalBreakpoint) {
+  @include whenNarrowerThan($guideNormalBreakpoint) {
     $leftMargin: 20px + $guideSideNavSmallWidth;
     $rightMargin: 20px;
