+++ ui_framework/doc_site/src/components/guide_page_side_nav/_guide_page_side_nav.scss
@@ -10,7 +10,7 @@
 
   @include scrollbar;
 
-  @include whenNarrowerThan($normalBreakpoint) {
+  @include whenNarrowerThan($guideNormalBreakpoint) {
     width: $guideSideNavSmallWidth;
   }
 }
