+++ ui_framework/doc_site/src/views/app_view.js
@@ -3,10 +3,6 @@ import React, {
   PropTypes,
 } from 'react';
 
-import {
-  Link,
-} from 'react-router';
-
 import classNames from 'classnames';
 
 import {
@@ -25,14 +21,8 @@ export class AppView extends Component {
   constructor(props) {
     super(props);
 
-    const currentRoute = props.routes[1];
-    const nextRoute = Routes.getNextRoute(currentRoute.name);
-    const previousRoute = Routes.getPreviousRoute(currentRoute.name);
-
     this.state = {
       isNavOpen: false,
-      nextRoute,
-      previousRoute,
     };
 
     this.onClickNavItem = this.onClickNavItem.bind(this);
@@ -56,52 +46,6 @@ export class AppView extends Component {
     });
   }
 
-  componentWillReceiveProps(nextProps) {
-    const currentRoute = nextProps.routes[1];
-    const nextRoute = Routes.getNextRoute(currentRoute.name);
-    const previousRoute = Routes.getPreviousRoute(currentRoute.name);
-
-    this.setState({
-      nextRoute,
-      previousRoute,
-    });
-  }
-
-  renderPagination() {
-    const previousClasses = classNames('guidePaginationButton', {
-      'guidePaginationButton-isDisabled': !this.state.previousRoute,
-    });
-
-    const previousButton = (
-      <Link
-        className={previousClasses}
-        to={this.state.previousRoute ? this.state.previousRoute.path : ''}
-      >
-        <span className="fa fa-angle-left"></span>
-      </Link>
-    );
-
-    const nextClasses = classNames('guidePaginationButton', {
-      'guidePaginationButton-isDisabled': !this.state.nextRoute,
-    });
-
-    const nextButton = (
-      <Link
-        className={nextClasses}
-        to={this.state.nextRoute ? this.state.nextRoute.path : ''}
-      >
-        <span className="fa fa-angle-right"></span>
-      </Link>
-    );
-
-    return (
-      <div className="guidePaginationButtons">
-        {previousButton}
-        {nextButton}
-      </div>
-    );
-  }
-
   render() {
     const contentClasses = classNames('guideContent', {
       'is-code-viewer-open': this.props.isCodeViewerOpen,
@@ -114,6 +58,9 @@ export class AppView extends Component {
           onToggleNav={this.onToggleNav}
           onClickNavItem={this.onClickNavItem}
           version={pkg.version}
+          routes={this.props.routes}
+          getNextRoute={Routes.getNextRoute}
+          getPreviousRoute={Routes.getPreviousRoute}
           components={Routes.components}
           sandboxes={Routes.sandboxes}
         />
@@ -128,8 +75,6 @@ export class AppView extends Component {
           title={this.props.title}
           source={this.props.source}
         />
-
-        {this.renderPagination()}
       </div>
     );
   }
